MILAB-6274: extract defaultBindingsFor into samplesheet-bindings helper
Pure refactor — no behavior change. Moves the binding-decision logic out of
the Vue component so it can be unit-tested in isolation. The characterization
test on origin/main behavior is the failing-test anchor for the bug fix in
the next commit.
